Add the data source health check subsystem
The patch will make it possible to check whether a Grafana data source
is healthy, in the same manner like the "Save & test" button works, when
creating a data source in the user interface.

It will support "server"- and "proxy"-type data source items.

The currently supported data sources are:

- CrateDB
- Elasticsearch
- InfluxDB
- PostgreSQL
- Prometheus
- Testdata